requirements.txt
Django==2.0.4
pytz==2018.3
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.4
google-api-python-client==1.6.6
httplib2==0.11.3
idna==2.6
oauth2client==4.1.2
oauthlib==2.0.7
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
uritemplate==3.0.0
urllib3==1.22
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.4
django-js-asset==1.0.0
django-tinymce==2.7.0
google-api-python-client==1.6.6
httplib2==0.11.3
idna==2.6
oauth2client==4.1.2
oauthlib==2.0.7
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
uritemplate==3.0.0
urllib3==1.22
